Moved extension files to extensions package.
