+++ composeApp/src/commonMain/kotlin/utils/extensions/BooleanExtensions.kt
@@ -1,0 +1,4 @@
+package utils.extensions
+
+fun Boolean.toLong(): Long =
+    if (this.not()) 0 else 1
