+++ composeApp/src/commonMain/kotlin/utils/extensions/StringExtensions.kt
@@ -1,4 +1,4 @@
-package utils
+package utils.extensions
 
 fun String.canBeLong() : Boolean {
     return this.toLongOrNull() != null
